Refactor link usage cron job and remove unused code
